2509 update 2

+++ RPD_API/Models/GrowthRate.cs
@@ -9,7 +9,7 @@ namespace RPD_API.Models
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public Guid growthRateID { get; set; }
         public string grName { get; set; }
-        public double grTotalExp { get; set; }
+        public int grTotalExp { get; set; }
 
         public ICollection<Pokemons> Pokemons { get; set; }
     }
